StorageLoader: Support skip_missing_jsonpaths option
If a shredded event exists in s3 out folder, but no jsonpath is defined
in iglu repos, setting skip_missing_jsonpaths to true will simply ignore
this event withou raising an error.

This is usefull when you do not what to store in Redshift all your Self
Described JSon events.