Scala Stream Collector: allow use of the originating scheme during cookie bounce (closes #3512)
Commit adds a new configuration parameter (`forwardedProtocolHeader`) in the
cookie bounce feature. This parameter specifies the name of an http header
that contains the originating protocol if deployed behind a load balancer.

This is useful if SSL termination happens at the load balancer in order to
maintain the security model of you web page.